% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{account}
\alias{account}
\title{Create an account object}
\usage{
account(user = Sys.getenv("SLUSER"), password = Sys.getenv("SLPASS"))
}
\arguments{
\item{user}{The SauceLabs user. By default an environmental variable "SLUSER" is looked for.}

\item{password}{The SauceLabs password for user. By default an environmental variable "SLPASS" is looked for.}
}
\value{
returns an object of class "account".
}
\description{
Creates a wrapper around the httr \code{\link{authenticate}} function. Adds an "account" class.
}
\examples{
\dontrun{
myAcc <- account()
}
}

