% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{changeAccessKey}
\alias{changeAccessKey}
\title{changeAccessKey}
\usage{
changeAccessKey(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
changeAccessKey (account =  , ... =  )
}

