% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{createUser}
\alias{createUser}
\title{Create a sub account}
\usage{
createUser(account, username = Sys.getenv("SLUSER"), newUsername, password,
  name, email, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{newUsername}{The username of the new user you wish to create}

\item{password}{The password for the new user you wish to create}

\item{name}{The name of the new user you wish to create}

\item{email}{The email of the new user you wish to create}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Create a sub account
}
\examples{
createUser (account =  , username =  , password =  , name =  , email =  , ... =  )
}

