% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMethods.R
\name{getAppiumEolDates}
\alias{getAppiumEolDates}
\title{getAppiumEolDates}
\usage{
getAppiumEolDates(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getAppiumEolDates (account =  , ... =  )
}

