% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMethods.R
\name{getJobAssetNames}
\alias{getJobAssetNames}
\title{getJobAssetNames}
\usage{
getJobAssetNames(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getJobAssetNames (account =  , ... =  )
}

