% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUnitTestMethods.R
\name{getJsUnitTestStatus}
\alias{getJsUnitTestStatus}
\title{Get JS Unit Test Status}
\usage{
getJsUnitTestStatus(account, username = Sys.getenv("SLUSER"), js_tests, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{...}{Additonal function arguments - Currently unused.}

\item{jsTests}{a vector of job ids.}
}
\description{
Get the status of your JS unit tests
}
\examples{
getJsUnitTestStatus (account =  , js_tests =  , ... =  )
}

