% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getListOfSiblingAccounts}
\alias{getListOfSiblingAccounts}
\title{Get a list of sibling accounts}
\usage{
getListOfSiblingAccounts(account, username = Sys.getenv("SLUSER"),
  page = NULL, per_page = 50L, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{page}{optional defaults to NULL}

\item{per_page}{results per page (max 50L). Defaults to 50L}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Get a list of sibling accounts associated with provided account
}
\examples{
getListOfSiblingAccounts (account =  , page =  , per_page =  , ... =  )
}

