% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actUsageMethods.R
\name{getRealTimeJobActivity}
\alias{getRealTimeJobActivity}
\title{getRealTimeJobActivity}
\usage{
getRealTimeJobActivity(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getRealTimeJobActivity (account =  , ... =  )
}

