% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getSubAccountInformation}
\alias{getSubAccountInformation}
\title{getSubAccountInformation}
\usage{
getSubAccountInformation(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getSubAccountInformation (account =  , ... =  )
}

