% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnelMethods.R
\name{getTunnel}
\alias{getTunnel}
\title{getTunnel}
\usage{
getTunnel(account, username = Sys.getenv("SLUSER"), tunnelID, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{tunnelID}{id of the tunnel to get more information on}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Get information for a tunnel given its ID
}
\examples{
getTunnel (account =  , ... =  )
}

