% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getUser}
\alias{getUser}
\title{Access basic account information}
\usage{
getUser(account, username = Sys.getenv("SLUSER"), ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Access basic account information
}
\examples{
getUser (account =  , ... =  )
}

