% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actUsageMethods.R
\name{getUserAccountUsage}
\alias{getUserAccountUsage}
\title{Get User Account Usage}
\usage{
getUserAccountUsage(account, username = Sys.getenv("SLUSER"), ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Access historical account usage data
}
\examples{
getUserAccountUsage (account =  , ... =  )
}

