% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getUserConcurrency}
\alias{getUserConcurrency}
\title{getUserConcurrency}
\usage{
getUserConcurrency(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getUserConcurrency (account =  , ... =  )
}

