% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{queryAPI}
\alias{queryAPI}
\title{Send a query to SauceLabs.}
\usage{
queryAPI(verb = GET, account, url, source, ...)
}
\arguments{
\item{verb}{The http method to use. See \code{\link{VERB}}}

\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{url}{The url of the remote server endpoint.}

\item{source}{The name of the RSauceLabs function that called queryDriver.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
The contents of the response from the remote server. See \code{\link{content}} for details.
}
\description{
\code{queryAPI} A function to send a query to SauceLabs. Intended for seleniumPipes
   internal use mainly.
}
\examples{
\dontrun{
# function intended for internal use
}
}

