% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUnitTestMethods.R
\name{startJsUnitTests}
\alias{startJsUnitTests}
\title{Start JS Unit Tests}
\usage{
startJsUnitTests(account, username = Sys.getenv("SLUSER"), platforms, url,
  framework, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{platforms}{A list of platforms (see example)}

\item{url}{should point to the page that hosts your tests}

\item{framework}{can be "qunit", "jasmine", "YUI Test", "mocha", or "custom"}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Start your JavaScript unit tests on as many browsers as you like with a single request
}
\examples{
startJsUnitTests (account =  , platforms =  , url =  , framework =  , ... =  )
}

