% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMethods.R
\name{updateJob}
\alias{updateJob}
\title{updateJob}
\usage{
updateJob(account, name, tags, public, passed, build, custom_data, ...)
}
\arguments{
\item{account}{}

\item{name}{}

\item{tags}{}

\item{public}{}

\item{passed}{}

\item{build}{}

\item{custom_data}{}

\item{...}{}
}
\examples{
updateJob (account =  , name =  , tags =  , public =  , passed =  , build =  , custom_data =  , ... =  )
}

