% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{createUser}
\alias{createUser}
\title{createUser}
\usage{
createUser(account, username, password, name, email, ...)
}
\arguments{
\item{account}{}

\item{username}{}

\item{password}{}

\item{name}{}

\item{email}{}

\item{...}{}
}
\examples{
createUser (account =  , username =  , password =  , name =  , email =  , ... =  )
}

