% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMethods.R
\name{deleteJob}
\alias{deleteJob}
\title{Delete Job}
\usage{
deleteJob(account, username = Sys.getenv("SLUSER"), jobID, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{jobID}{Id of the job to delete}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Removes the job from the system with all the linked assets
}
\examples{
deleteJob (account =  , ... =  )
}

