% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMethods.R
\name{deleteJobAssets}
\alias{deleteJobAssets}
\title{deleteJobAssets}
\usage{
deleteJobAssets(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
deleteJobAssets (account =  , ... =  )
}

