% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMethods.R
\name{getAppiumEolDates}
\alias{getAppiumEolDates}
\title{Get Appium EOL dates}
\usage{
getAppiumEolDates(account, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{...}{Additonal function arguments - Currently unused.}

\item{username}{SauceLabs username}
}
\description{
Get a list of Appium end-of-life dates. Dates are displayed in Unix time.
}
\examples{
getAppiumEolDates (account =  , ... =  )
}

