% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMethods.R
\name{getAppiumEolDates}
\alias{getAppiumEolDates}
\title{getAppiumEolDates}
\usage{
getAppiumEolDates(account, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{...}{Additonal function arguments - Currently unused.}

\item{username}{SauceLabs username}
}
\examples{
getAppiumEolDates (account =  , ... =  )
}

