% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMethods.R
\name{getJobAssetFiles}
\alias{getJobAssetFiles}
\title{getJobAssetFiles}
\usage{
getJobAssetFiles(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getJobAssetFiles (account =  , ... =  )
}

