% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMethods.R
\name{getJobs}
\alias{getJobs}
\title{Get Jobs}
\usage{
getJobs(account, username = Sys.getenv("SLUSER"), limit = 100L,
  getFullJobs = FALSE, skipJobs = 0L, to = NULL, from = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
List recent jobs belonging to a specific user
}

