% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUnitTestMethods.R
\name{getJsUnitTestStatus}
\alias{getJsUnitTestStatus}
\title{getJsUnitTestStatus}
\usage{
getJsUnitTestStatus(account, js_tests, ...)
}
\arguments{
\item{account}{}

\item{js_tests}{}

\item{...}{}
}
\examples{
getJsUnitTestStatus (account =  , js_tests =  , ... =  )
}

