% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getListOfSiblingAccounts}
\alias{getListOfSiblingAccounts}
\title{getListOfSiblingAccounts}
\usage{
getListOfSiblingAccounts(account, page, per_page, ...)
}
\arguments{
\item{account}{}

\item{page}{}

\item{per_page}{}

\item{...}{}
}
\examples{
getListOfSiblingAccounts (account =  , page =  , per_page =  , ... =  )
}

