% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMethods.R
\name{getSauceLabsStatus}
\alias{getSauceLabsStatus}
\title{getSauceLabsStatus}
\usage{
getSauceLabsStatus(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getSauceLabsStatus (account =  , ... =  )
}

