% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMethods.R
\name{getSauceLabsStatus}
\alias{getSauceLabsStatus}
\title{Get Sauce Labs Status}
\usage{
getSauceLabsStatus(account, ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Get the current status of Sauce Labs services
}
\examples{
getSauceLabsStatus (account =  , ... =  )
}

