% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempStorageMethods.R
\name{getStoredFiles}
\alias{getStoredFiles}
\title{getStoredFiles}
\usage{
getStoredFiles(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getStoredFiles (account =  , ... =  )
}

