% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getSubAccountInformation}
\alias{getSubAccountInformation}
\title{Get information about a sub account}
\usage{
getSubAccountInformation(account, username = Sys.getenv("SLUSER"), ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Get information about a sub account
}
\examples{
getSubAccountInformation (account =  , ... =  )
}

