% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMethods.R
\name{getSupportedPlatforms}
\alias{getSupportedPlatforms}
\title{getSupportedPlatforms}
\usage{
getSupportedPlatforms(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getSupportedPlatforms (account =  , ... =  )
}

