% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnelMethods.R
\name{getTunnels}
\alias{getTunnels}
\title{Get Tunnels}
\usage{
getTunnels(account, username = Sys.getenv("SLUSER"), ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Retrieves all running tunnels for a specific user
}
\examples{
getTunnels (account =  , ... =  )
}

