% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actUsageMethods.R
\name{getUserAccountUsage}
\alias{getUserAccountUsage}
\title{getUserAccountUsage}
\usage{
getUserAccountUsage(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getUserAccountUsage (account =  , ... =  )
}

