% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actUsageMethods.R
\name{getUserActivity}
\alias{getUserActivity}
\title{getUserActivity}
\usage{
getUserActivity(account, ...)
}
\arguments{
\item{account}{}

\item{...}{}
}
\examples{
getUserActivity (account =  , ... =  )
}

