% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountMethods.R
\name{getUserConcurrency}
\alias{getUserConcurrency}
\title{Check account concurrency limits}
\usage{
getUserConcurrency(account, username = Sys.getenv("SLUSER"), ...)
}
\arguments{
\item{account}{An object of class "account". An account object see \code{\link{account}}.}

\item{username}{SauceLabs username}

\item{...}{Additonal function arguments - Currently unused.}
}
\description{
Check account concurrency limits
}
\examples{
getUserConcurrency (account =  , ... =  )
}

