% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUnitTestMethods.R
\name{startJsUnitTests}
\alias{startJsUnitTests}
\title{startJsUnitTests}
\usage{
startJsUnitTests(account, platforms, url, framework, ...)
}
\arguments{
\item{account}{}

\item{platforms}{}

\item{url}{}

\item{framework}{}

\item{...}{}
}
\examples{
startJsUnitTests (account =  , platforms =  , url =  , framework =  , ... =  )
}

